cpio: work on other GOARCHes

+++ cpio/sysinfo_linux.go
@@ -9,19 +9,20 @@ import (
 )
 
 func sysInfo(n string, sys *syscall.Stat_t) Info {
+	//nolint:unconvert
 	return Info{
 		Ino:      sys.Ino,
 		Mode:     uint64(sys.Mode),
 		UID:      uint64(sys.Uid),
 		GID:      uint64(sys.Gid),
-		NLink:    sys.Nlink,
+		NLink:    uint64(sys.Nlink),
 		MTime:    uint64(sys.Mtim.Sec),
 		FileSize: uint64(sys.Size),
-		Dev:      sys.Dev,
-		Major:    sys.Dev >> 8,
-		Minor:    sys.Dev & 0xff,
-		Rmajor:   sys.Rdev >> 8,
-		Rminor:   sys.Rdev & 0xff,
+		Dev:      uint64(sys.Dev),
+		Major:    uint64(sys.Dev >> 8),
+		Minor:    uint64(sys.Dev & 0xff),
+		Rmajor:   uint64(sys.Rdev >> 8),
+		Rminor:   uint64(sys.Rdev & 0xff),
 		Name:     n,
 	}
 }
